Base refactoring done, needs further improvement in other areas for consistency
I restructured it in a way that is more verbose but the point is that everyone reading it should understand which variable belongs to which object or item.

It would be nice to follow the same structure for consistency when creating new objects:
-plateWars.object
-- plateWars.object.baseData, config, records, refIds and so on.

It's much easier to look something up or reference it within the code.

TODO:
What should be done is creating functions for logging: LogInfo, LogWarning, LogError
Create structure for messages, e.g. plateWars.messages
It will help in creating multiple sessions per server if we have basic stuff organized.